requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
python-pptx>=0.6.21
matplotlib>=3.7.0
scipy>=1.11.0
openai>=1.0.0
jsonschema>=4.23.0
